requirements.txt
graphics.py==5.0.1.post1
PyAudio==0.2.11
